Env,Nix: Add nix flake environment
This doesn't need to be "THE official environment", but it's
to be able to cd into the project and have the right tools to
work on it.
Feel free to push back if you think this is better maintained
on a personal environment. I can still keep it around untracked.

This adds a bunch of tools,
- rust (rustc, cargo, mold, ...)
- bacon
- rust-analyzer
- random rust/cargo tools that I sometimes end up reaching to,
  - cargo bloat
  - cargo deny
  - cargo edit
  - cargo modules
  - cargo outdated

And generates some pre-commit checks,
- Rust checks
  - cargo check
  - clippy
- Formatting (nix, toml, rust)
- Spelling (I picked en_GB)