Lint: Add a Github Action for running clippy.
This is not integrated with the nix flake environment, but its
still a good start given how easier it is.

With nix flakes it'll need some more testing as flakes support
architectures that Github Actions doesn't, like ARM.

BTW, I'm renaming the "rust" action to "release" since it's
clearer and more abstract.