BREAKING,Types: Use IriString instead of String to handle IRIs
There's a conversion layer to interact with reqwest, which only
handles URLs via `url::Url`, but we only convert right before
calling reqwest.

This breaks the history database ($XDG_DATA_HOME/rawst/history.json or ~/.local/share/rawst/history.json).

I think breaking that file is acceptable for now since we are pre-1.0.
The file can be removed, or can be patched by renaming the `url` key to `iri`.

We could workaround the breakage by telling serde to use "url" as the key
for the `iri` field, but I'd rather keep this a breaking change since
I think the breakage is not that bad.