Types: Use PathBuf instead of String
I'm not sure if the behaviour is still the same.
I guess we need to document what's the intended behaviour first.

When downloading a new file

Where does the file go?
  - rawst's `$PWD`?
  - `~/Downloads/` or its XDG equivalent?

How is the file named?
  - Given by link
  - Guessed by address
  - Specified by user

Where are the temporary chunks stored?
  - We use XDG_CACHE_HOME, but I think it's not a good directory.
    - It could be in a different filesystem
      - Weird stuff can happen regarding filesystem details
        - Name changes/clashes
        - File size limitations
        - Metadata thrown away or altered unintentionally
      - Remaining disk space and quota may be a problem.

I didn't focus on retaining the current behaviour, but snuck a bunch of asserts
that may help doing that. I'd rather spend time defining the intended behaviour
and thinking about how to maintain tests for that. I'm not familiar with testing
in Rust yet and testing entire programs that mutate the filesystem is likely a
bit tricky.