CI: Unify tests using nix flake environment.
This switches the build to a single run using just the nix flake environment. It
should be faster and more efficient. I don't think rawst needs platform specific
code, so testing it on multiple platforms seems like an overkill.

The release workflow remains, but I'm switching it to Rust nightly too.