feat!: rework to transfer httptask instead of Arguments
This commit replaces the transferring of Arguments to transferring of HttpTask instead. This was done so it allows to have a progressbar at the client side as well as on the server side and updates it in sync. WARNING: This is not finalised as this commit only implements the download function and not others.